+++ src/senaite/core/schema/__init__.py
@@ -9,6 +9,8 @@ from .fields import IntField
 from .interfaces import IDatetimeField
 from .interfaces import IIntField
 from .interfaces import IRichTextField
+from .phonefield import IPhoneField
+from .phonefield import PhoneField
 from .richtextfield import RichTextField
 from .uidreferencefield import IUIDReferenceField
 from .uidreferencefield import UIDReferenceField
@@ -16,5 +18,6 @@ from .uidreferencefield import UIDReferenceField
 classImplementsFirst(AddressField, IAddressField)
 classImplementsFirst(DatetimeField, IDatetimeField)
 classImplementsFirst(IntField, IIntField)
+classImplementsFirst(PhoneField, IPhoneField)
 classImplementsFirst(RichTextField, IRichTextField)
 classImplementsFirst(UIDReferenceField, IUIDReferenceField)