+++ src/senaite/core/z3cform/static/datagrid.js
@@ -188,6 +188,8 @@ document.addEventListener("DOMContentLoaded", () => {
       row.remove();
       this.update_order_index(datagrid);
       this.init_auto_append_handler(datagrid);
+      let event_data = {datagrid: datagrid, row: row}
+      this.trigger_custom_event("datagrid:row_removed", event_data);
     }
 
     insert_row(ref_row, before=false) {
@@ -200,8 +202,9 @@ document.addEventListener("DOMContentLoaded", () => {
       } else {
         $newtr.insertAfter(ref_row);
       }
-
       this.init_auto_append_handler(datagrid);
+      let event_data = {datagrid: datagrid, row: newtr}
+      this.trigger_custom_event("datagrid:row_added", event_data);
       return newtr;
     }
 
@@ -355,6 +358,11 @@ document.addEventListener("DOMContentLoaded", () => {
           }
         }
       }
+    }
+
+    trigger_custom_event(event_name, event_data) {
+      let event = new CustomEvent(event_name, {detail: event_data});
+      document.body.dispatchEvent(event);
     }
 
     /* EVENT HANDLERS */